airflowDags/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow==2.7.1
astro-runtime-python==6.0.0
astro-sdk-python==1.7.0
requests>=2.31.0
boto3>=1.29.3
pandas>=2.1.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
